Fixed Alignment logic errors
Fixed an error after changing to class in FindCheckpoint class - caused by == operator and my human error
Fixed a logic error for the alignment. The distance to next is calculated after the next checkpoint, so i can't use it in the first loop. so I might as well calculate it.